add "placeholder" for calculate interest
we should add interest from each utxo to sum of inputs, but the
issue that we haven't access to prevout tx here to get locktime.
we have only hash and index for prevout. so, we should somehow
get the tx, like we did in komodo_accrued_interest/komodo_interest_args
with GetTransaction.